Toggle resource is template in resource header
